requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
g4f==0.1.9.5 
